Clean up lint/vet issues and potential races (#289)
* Fix go vet nits

* megacheck lint issues

* Resolve race condition issues, and mark a few tests as exempt from race detection

* gofmt fixes

* lint fixes for package godocs
