+++ core/internal/doc.go
@@ -1,4 +1,4 @@
-// Here be dragons.
+// Package internal - Here be dragons.
 // The internal package contains the bulk of Burrow's logic, including all of the coordinators and modules that have
 // been defined. This documentation is targeted at developers of Burrow modules to get more information about the
 // internal structure and how the modules fit together. It is not designed for end users, and for the most part will not
