+++ core/internal/notifier/http.go
@@ -39,7 +39,6 @@ type HTTPNotifier struct {
 	Log *zap.Logger
 
 	name           string
-	threshold      int
 	groupWhitelist *regexp.Regexp
 	groupBlacklist *regexp.Regexp
 	extras         map[string]string
