+++ core/internal/notifier/null.go
@@ -31,7 +31,6 @@ type NullNotifier struct {
 	Log *zap.Logger
 
 	name           string
-	threshold      int
 	groupWhitelist *regexp.Regexp
 	groupBlacklist *regexp.Regexp
 	extras         map[string]string
